backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow>=2.15.0
scikit-learn
joblib
numpy
pillow
opencv-python-headless
scikit-image
xgboost
